split up components
